% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapted_discrepancy_continuous_sum.R
\name{adapted_discrepancy_continuous_sum}
\alias{adapted_discrepancy_continuous_sum}
\title{Adapted disc_func for EEM}
\usage{
adapted_discrepancy_continuous_sum(
  data,
  target_lower = NULL,
  target_upper = NULL
)
}
\arguments{
\item{data}{vector of the equilibrium points followed by the real eigenvalues of the jacobian matrix, as returned by \link[EEMtoolbox]{summarise_ecosystem_features}}

\item{target_lower}{vector of the lower bound of the target equilibrium}

\item{target_upper}{vector of the upper bound of the target equilibrium}
}
\value{
summary statistic (discrepancy measure).
}
\description{
Discrepancy measure: summarises how infeasible or instable this system is + adds penalty for wrong equilibria
}
