% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dingo_matrix}
\alias{dingo_matrix}
\title{Dingo interaction matrix}
\format{
A matrix with 8 rows and 8 columns
\describe{
\item{Dingo}{Dingo positive or negative influence on other nodes}
\item{Mesopredators}{Mesopredators positive or negative influence on other nodes}
\item{Large Herbivores}{Large Herbivores: positive or negative influence on other nodes}
\item{Small Vertebrates}{Small Herbivores: positive or negative influence on other nodes}
\item{Grass}{Grass: positive or negative influence on other nodes}
\item{Invertebrates}{Invertebrates: positive or negative influence on other nodes}
\item{Fire}{Fire: positive or negative influence on other nodes}
\item{Soil}{Soil: positive or negative influence on other nodes}
}
}
\usage{
dingo_matrix
}
\description{
Dingo interaction matrix
}
\keyword{datasets}
