% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrepancy_continuous_sum.R
\name{discrepancy_continuous_sum}
\alias{discrepancy_continuous_sum}
\title{Discrepancy function}
\usage{
discrepancy_continuous_sum(data)
}
\arguments{
\item{data}{vector of the equilibrium points followed by the real eigenvalues of the jacobian matrix, as returned by \link[EEMtoolbox]{summarise_ecosystem_features}}
}
\value{
summary statistic (discrepancy measure).
}
\description{
Discrepancy measure: summarises how infeasible or instable this system is
}
