% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EEM_SMC_method.R
\name{EEM_SMC_method}
\alias{EEM_SMC_method}
\title{Generation of model ensembles Sequential Monte Carlo - Approximate Bayesian Computation.}
\usage{
EEM_SMC_method(
  sim_args,
  summ_func,
  disc_func,
  sampler,
  trans_f,
  trans_finv,
  pdf,
  n_particles,
  mcmc_trials,
  dist_final,
  a,
  c,
  p_acc_min
)
}
\arguments{
\item{sim_args}{a list of arguments as returned by \link[EEMtoolbox]{args_function}}

\item{summ_func}{function calculating equilibrium points and real parts of the Jacobians eigenvalues to summarise ecosystem features.}

\item{disc_func}{summary statistic (discrepancy measure).}

\item{sampler}{sampling function that generates random vectors from the joint prior distribution.}

\item{trans_f}{transform of prior parameter space to ensure unbounded support for MCMC sampling.}

\item{trans_finv}{inverse of trans_f function.}

\item{pdf}{joint probability density function.}

\item{n_particles}{number of particles in the sample.}

\item{mcmc_trials}{number of MCMC steps to try before selecting appropriate number.}

\item{dist_final}{target discrepancy threshold. Default 0. If zero, p_acc_min is used to determine stopping criteria.}

\item{a}{tuning parameter for adaptive selection of discrepancy threshold sequence.}

\item{c}{tuning parameter for choosing the number of MCMC iterations in move step.}

\item{p_acc_min}{minimum acceptable acceptance rate in the MCMC interations before exit.}
}
\value{
list: sims=number of simulations
part_vals=parameter values
part_s=discrepancy value
prior_sample=prior distribution
}
\description{
Generation of model ensembles using Adaptive sequential Monte Carlo for approximate Bayesian computation
}
