% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.R
\name{MCMC}
\alias{MCMC}
\title{Monte Carlo Markov Chain}
\usage{
MCMC(
  i,
  sim_args,
  mcmc_trials,
  dist_next,
  part_vals,
  part_s,
  part_sim,
  cov_matrix,
  summ_func,
  disc_func,
  trans_finv,
  pdf
)
}
\arguments{
\item{i}{line number of particle to be moved}

\item{sim_args}{a list of arguments as returned by \link[EEMtoolbox]{args_function}}

\item{mcmc_trials}{number of MCMC steps to try before selecting appropriate number.}

\item{part_vals}{matrix of current values of parameters of each particle (each particle represented by a row)}

\item{part_s}{vector of discrepancy measures of each particle}

\item{part_sim}{(each particle represented by a row)}

\item{cov_matrix}{covariance matrix}

\item{summ_func}{function calculating equilibrium points and real parts of the Jacobians eigenvalues to summarise ecosystem features. Default =summarise_ecosystem_features_GLV. Options include summarise_ecosystem_features_Baker (automatically chosen if model="Baker") and summarise_ecosystem_features_Adams, (automatically chosen if model="Adams"). Needs to be defined if model="customized" chosen.}

\item{disc_func}{summary statistic (discrepancy measure).}

\item{trans_finv}{inverse of trans_f function.}

\item{pdf}{joint probability density function.}
}
\value{
list
part_vals: updated value of parameter values for particle i
part_s: discrepancy measure for particle i
part_sim
i_acc
sims_mcmc
}
\description{
Runs MCMC
}
