% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_cores_function.R
\name{n_cores_function}
\alias{n_cores_function}
\title{Number of cores}
\usage{
n_cores_function(n_cores = 1L)
}
\arguments{
\item{n_cores}{Number of cores desired for parallel sampling. Default set to 1 core (sequential sampling).}
}
\description{
Gets number of cores available for parallel tasks
}
