% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_species_function.R
\name{n_species_function}
\alias{n_species_function}
\title{Arguments for EEM}
\usage{
n_species_function(interaction_matrix, model = "GLV")
}
\arguments{
\item{interaction_matrix}{interaction signs matrix, can be input as a single matrix of interactions or as a list of matrices defining lower and upper bounds for interaction terms lower first and upper second}

\item{model}{model representing species interactions, default "GLV" (Generalized Lotka Volterra). options include "Bimler-Baker", "Gompertz" and "customized"}
}
\value{
A list of arguments defining the problem.

n_species: number of species
}
\description{
Extract number of species from interaction matrix
}
\examples{
n_species_function(dingo_matrix) #automatically loads an example of interaction matrix as dingo_matrix
}
