% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_solve.R
\name{ode_solve}
\alias{ode_solve}
\title{Derivatives of models}
\usage{
ode_solve(parameters, t_window, IC)
}
\arguments{
\item{interaction_matrix}{interaction signs matrix, can be input as a single matrix of interactions or as a list of matrices defining lower and upper bounds for interaction terms lower first and upper second}

\item{bounds_growth_rate}{vector of growth rates}

\item{current_abundance}{vector of current species abundance}

\item{model}{model representing species interactions. Default "GLV" (Generalized Lokta Voltera). options include "Baker", "Gompertz" and "customized"}
}
\value{
list: part_vals: ensemble of parameters, marginal distributions
}
\description{
Derivatives functions to solve ODE systems
}
\examples{
library(EEMtoolbox)
EEM(dingo_matrix) #automatically loads an example of interaction matrix as dingo_matrix
}
