% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{projections}
\alias{projections}
\title{Solve ODE}
\usage{
projections(
  parameters,
  initial_condition,
  t_window,
  time_step_len = 0.01,
  model = "GLV",
  derivative = EEMtoolbox::derivative_func
)
}
\arguments{
\item{parameters}{list like object of ensemble of parameters (outputs of  \link[EEMtoolbox]{EEM})}

\item{initial_condition}{vector of initial species abundances}

\item{t_window}{time window to solve ODE}

\item{time_step_len}{length of each time step, default = 0.01}

\item{model}{model representing species interactions. Default "GLV" (Generalized Lotka Volterra). options include "Baker" and "Gompertz".}

\item{derivative}{derivative function. Default \link[EEMtoolbox]{derivative_func}}
}
\value{
matrix: first column (time) vector of time steps of the simulation, the following columns correspond to the values of abundances at a given time step
}
\description{
Solves ODEs
}
\examples{
library(EEMtoolbox)
output <- EEM(dingo_matrix) #automatically loads an example of interaction matrix as dingo_matrix
projections(output, t_window=c(0,1))
}
