% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_matrix_growthrates.R
\name{reconstruct_matrix_growthrates}
\alias{reconstruct_matrix_growthrates}
\title{Reconstruct interaction matrix and growth rates vector from set of parameters}
\usage{
reconstruct_matrix_growthrates(parameters, sim_args)
}
\arguments{
\item{parameters}{a vector of sampled parameters}

\item{sim_args}{a list of arguments as returned by \link[EEMtoolbox]{args_function}}
}
\value{
A list of 2 elements
growthrates: a vector of arguments defining the problem,
interaction_matrix: reconstructed interaction matrix.
If model is Bimler-Baker,
interaction_matrix_alphas: reconstructed interaction matrix of alphas
interaction_matrix_betas: reconstructed interaction matrix of betas
}
\description{
Reconstruct interaction matrix and growth rates vector from set of parameters
}
