% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_matrix_growthrates_Baker.R
\name{reconstruct_matrix_growthrates_Baker}
\alias{reconstruct_matrix_growthrates_Baker}
\title{Reconstruct interaction matrix and growth rates vector from set of parameters}
\usage{
reconstruct_matrix_growthrates_Baker(parameters, sim_args)
}
\arguments{
\item{parameters}{a vector of sampled parameters}

\item{sim_args}{a list of arguments as returned by \link[EEMtoolbox]{args_function}}
}
\value{
A list of 3 elements growthrates: a vector of arguments defining the problem. alphas_matrix=matrix of alpha elements and interaction_matrix: reconstructed interaction matrix
}
\description{
Reconstruct interaction matrix and growth rates vector from set of parameters
}
