% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_list}
\alias{species_list}
\title{List of species}
\format{
A vector with the species names
\describe{
\item{Dingo}{Dingo positive or negative influence on other nodes}
\item{Mesopredators}{Mesopredators positive or negative influence on other nodes}
\item{Large Herbivores}{Large Herbivores: positive or negative influence on other nodes}
\item{Small Vertebrates}{Small Herbivores: positive or negative influence on other nodes}
\item{Grass}{Grass: positive or negative influence on other nodes}
\item{Invertebrates}{Invertebrates: positive or negative influence on other nodes}
\item{Fire}{Fire: positive or negative influence on other nodes}
\item{Soil}{Soil: positive or negative influence on other nodes}
}
}
\usage{
species_list
}
\description{
List of species
}
\keyword{datasets}
