% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_sampler.R
\name{uniform_sampler}
\alias{uniform_sampler}
\title{Uniform sampling of parameters}
\usage{
uniform_sampler(sim_args)
}
\arguments{
\item{sim_args}{a list of arguments as returned by \link[EEMtoolbox]{args_function}}
}
\value{
a vector of sampled parameters
}
\description{
Uniform sampling of parameters between lower and upper bounds defined in sim_args
}
