% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_transform.R
\name{uniform_transform}
\alias{uniform_transform}
\title{transformation function}
\usage{
uniform_transform(theta, sim_args)
}
\arguments{
\item{theta}{vector of parameters to transform}

\item{sim_args}{a list of arguments as returned by \link[EEMtoolbox]{args_function}}
}
\value{
vector of transformed parameters
}
\description{
This function uses a transform for uniform distributions, so all values are within the uniform bounds.
}
