% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_transform_inverse.R
\name{uniform_transform_inverse}
\alias{uniform_transform_inverse}
\title{inverse transformation function}
\usage{
uniform_transform_inverse(theta_trans, sim_args)
}
\arguments{
\item{theta_trans}{vector of parameters to transformed back}

\item{sim_args}{a list of arguments as returned by \link[EEMtoolbox]{args_function}}
}
\value{
vector of transformed parameters
}
\description{
This function uses a reverts the transformation for uniform distributions
}
