% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args_function.R
\name{args_function}
\alias{args_function}
\title{Arguments for EEM}
\usage{
args_function(
  interaction_matrix,
  upper_bounds_growth_rate = 5,
  lower_bounds_growth_rate = 0,
  upper_interaction_strength = 1,
  model = "GLV"
)
}
\arguments{
\item{interaction_matrix}{interaction signs matrix. If model is GLV or Gompertz it can be input as a single matrix of interactions or as a list of matrices defining lower and upper bounds for interaction terms lower first and upper second.     #if model is Baker, the interaction_matrix has to be a list of two lists, the first list contains matrices defining lower and upper bounds of alphas, the second list contains matrices defining lower and upper bounds of betas}

\item{upper_bounds_growth_rate}{upper bound of growth rates. Input can be one number (same upper bound for all species) or a vector of growth rates upper bounds for each species. Default 5}

\item{lower_bounds_growth_rate}{lower bound of growth rates. Input can be one number (same lower bound for all species) or a vector of growth rates lower bounds for each species. Default 0}

\item{upper_interaction_strength}{coefficient representing interaction strength. Default:1}

\item{model}{model representing species interactions, default "GLV" (Generalized Lotka Volterra). options include "Baker", "Adams" and "customized"}
}
\value{
A list of arguments defining the problem.

list:
model: considered model
n_species: number of species
keep_parameters: vector of 0-1 indicating which parameters to keep
num_params: number of non-zero parameters
lower: vector of length num_params representing lower bound of each parameter
upper: vector of length num_params representing upper bound of each parameter
}
\description{
Extract arguments necessary to run EEM from interaction matrix
}
\examples{
library(EEMtoolbox)
args_function(dingo_matrix) #automatically loads an example of interaction matrix as dingo_matrix
}
