% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args_function.R
\name{args_function}
\alias{args_function}
\title{Arguments for EEM}
\usage{
args_function(
  interaction_matrix,
  bounds_growth_rate,
  upper_interaction_strength = 1,
  model = "GLV"
)
}
\arguments{
\item{interaction_matrix}{interaction signs matrix, can be input as a single matrix of interactions or as a list of matrices defining lower and upper bounds for interaction terms lower first and upper second}

\item{bounds_growth_rate}{vector of 2 elements containing lower and upper bounds for growth rates}

\item{upper_interaction_strength}{vector of 2 elements containing lower and upper bounds for growth rates}

\item{model}{model representing species interactions, default "GLV" (Generalized Lokta Voltera). options include "Baker", "Adams" and "customized"}
}
\value{
A list of arguments defining the problem.
}
\description{
Extract arguments necessary to run EEM from interaction matrix
}
