% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivative_func.R
\name{derivative_func}
\alias{derivative_func}
\title{Derivatives of models}
\usage{
derivative_func(Time, State, Pars)
}
\arguments{
\item{Time}{a float of current studied time}

\item{State}{vector of current species abundance}

\item{Pars}{a list like object containing: model: string of model representing species interactions. Options include "GLV" (Generalized Lotka Volterra), "Baker", "Gompertz" and "customized" interaction_matrix_value interaction matrix parametrized. If model is GLV or Gompertz, input a single matrix. If model is Baker, input a list of 2 matrices: the first element is the alphas and the second the betas., model representing species interactions. Default "GLV" (Generalized Lokta Voltera). Options include "Baker", "Gompertz". growth_rate: vector of growth rates}
}
\value{
list: part_vals: ensemble of parameters, marginal distributions
}
\description{
Derivatives functions to solve ODE systems
}
\examples{
library(EEMtoolbox)
output <- EEM(dingo_matrix) #automatically loads an example of interaction matrix as dingo_matrix
interaction_matrix_value <- output[[1]]$interaction_matrix
growth_rate <- output[[1]]$growthrates
current_abundance <- rep(10, 8) # 8 species, initial abundance is 10 for all
derivative_func(interaction_matrix_value, growth_rate, current_abundance)
}
