% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivative_func.R
\name{derivative_func}
\alias{derivative_func}
\title{Derivatives of models}
\usage{
derivative_func(
  interaction_matrix,
  growth_rate,
  current_abundance,
  model = "GLV"
)
}
\arguments{
\item{interaction_matrix}{interaction signs matrix, can be input as a single matrix of interactions or as a list of matrices defining lower and upper bounds for interaction terms lower first and upper second}

\item{current_abundance}{vector of current species abundance}

\item{model}{model representing species interactions. Default "GLV" (Generalized Lokta Voltera). options include "Baker", "Gompertz" and "customized"}

\item{bounds_growth_rate}{vector of growth rates}
}
\value{
list: part_vals: ensemble of parameters, marginal distributions
}
\description{
Derivatives functions to solve ODE systems
}
\examples{
library(EEMtoolbox)
EEM(dingo_matrix) #automatically loads an example of interaction matrix as dingo_matrix
}
