% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nonzero_parameters.R
\name{get_nonzero_parameters}
\alias{get_nonzero_parameters}
\title{List of non-zero parameters}
\usage{
get_nonzero_parameters(interaction_matrix)
}
\arguments{
\item{interaction_matrix}{interaction signs matrix, can be input as a single matrix of interactions or as a list of matrices defining lower and upper bounds for interaction terms lower first and upper second}
}
\value{
A list
skip_parameters : parameter skipped (yes=1) or not (no=0)
lower_interaction_bound: lower bounds of each parameter sign
upper_interaction_bound: upper bounds of each parameter sign
}
\description{
Array of indicating if a parameter in the interaction matrix is skipped (yes=1) or not (no=0), as well as the interaction sign (upper and lower bounds)
}
