% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_solve.R
\name{ode_solve}
\alias{ode_solve}
\title{Solve ODE}
\usage{
ode_solve(interaction_matrix, t_window, initial_condition, model = "GLV")
}
\arguments{
\item{interaction_matrix}{interaction signs matrix, can be input as a single matrix of interactions or as a list of matrices defining lower and upper bounds for interaction terms lower first and upper second}

\item{t_window}{time window to solve ODE}

\item{initial_condition}{vector of initial species abundances}

\item{model}{model representing species interactions. Default "GLV" (Generalized Lokta Voltera). options include "Baker", "Gompertz" and "customized"}
}
\value{
list: part_vals: ensemble of parameters, marginal distributions
}
\description{
Solves ODEs
}
\examples{
library(EEMtoolbox)
EEM(dingo_matrix) #automatically loads an example of interaction matrix as dingo_matrix
}
