% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_posterior_density.R
\name{plots_posterior_density}
\alias{plots_posterior_density}
\title{Arguments for EEM}
\usage{
plots_posterior_density(prior_sample, posterior_sample, sim_args, param_names)
}
\arguments{
\item{prior_sample}{interaction signs matrix, can be input as a single matrix of interactions or as a list of matrices defining lower and upper bounds for interaction terms lower first and upper second}

\item{posterior_sample}{vector of 2 elements containing lower and upper bounds for growth rates}

\item{sim_args}{vector of 2 elements containing lower and upper bounds for growth rates}

\item{param_names}{model representing species interactions, default "GLV" (Generalized Lokta Voltera). options include "Baker", "Adams" and "customized"}
}
\value{
A list of arguments defining the problem.
}
\description{
Extract arguments necessary to run EEM from interaction matrix
}
